requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.9
langchain-huggingface==0.0.3
huggingface-hub==0.20.3
sentence-transformers==2.3.1
faiss-cpu==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pypdf==3.17.4
jinja2==3.1.2
aiofiles==23.2.1
torch==2.1.2
transformers==4.36.2
